requirements.txt
asgiref==3.6.0
Django==4.1.5
et-xmlfile==1.1.0
mysqlclient==2.1.1
numpy==1.24.2
openpyxl==3.1.1
pandas==1.5.3
Pillow==9.4.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.7.1
qrcode==7.3.1
six==1.16.0
sqlparse==0.4.3
whitenoise==6.3.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
